# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sure-Bet-Analyzer API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type EventStatus = typeof EventStatus[keyof typeof EventStatus];


export const EventStatus = {
  pending: 'pending',
  live: 'live',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type Outcome = typeof Outcome[keyof typeof Outcome];


export const Outcome = {
  home: 'home',
  draw: 'draw',
  away: 'away',
} as const;

export interface Sport {
  id: number;
  name: string;
  slug: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreateSportRequest {
  name: string;
  slug: string;
  isActive?: boolean;
}

export interface Bookmaker {
  id: number;
  name: string;
  url?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateBookmakerRequest {
  name: string;
  url?: string;
  isActive?: boolean;
}

export interface Event {
  id: number;
  sportId: number;
  homeTeam: string;
  awayTeam: string;
  startsAt: string;
  status: EventStatus;
  createdAt: string;
}

export interface CreateEventRequest {
  sportId: number;
  homeTeam: string;
  awayTeam: string;
  startsAt: string;
  status?: EventStatus;
}

export interface UpdateEventStatusRequest {
  status: EventStatus;
}

export interface Odds {
  id: number;
  eventId: number;
  bookmakerId: number;
  outcome: Outcome;
  decimalOdds: number;
  updatedAt: string;
}

export interface UpsertOddsRequest {
  eventId: number;
  bookmakerId: number;
  outcome: Outcome;
  decimalOdds: number;
}

export interface StakeDetail {
  bookmakerId: number;
  odds: number;
  stake: number;
  payout: number;
}

export interface OpportunityStakes {
  home: StakeDetail;
  draw?: StakeDetail;
  away: StakeDetail;
}

export interface Opportunity {
  id: number;
  eventId: number;
  profitPercentage: number;
  stakes: OpportunityStakes;
  calculatedAt: string;
}

export interface ArbitrageCalculationResult {
  /** Number of new opportunities detected */
  found: number;
  opportunities: Opportunity[];
}

export type ListEventsParams = {
sportId?: number;
status?: EventStatus;
};

export type ListOddsParams = {
eventId?: number;
bookmakerId?: number;
};

export type ListOpportunitiesParams = {
eventId?: number;
/**
 * Minimum profit percentage (e.g. 1.5 means 1.5%)
 */
minProfit?: number;
};